Camera2: Inverted preview in Panorama
For 8x26MTP device, panorama preview is coming as inverted.
In present android version flippreview is itself taking care of
rotation. This additional rotation is causing the inverted preview
for the device where mount angle for sensor is more than 180degree.
The logic related to mPanoAngle is not needed which had been
added for previous android version.

Change-Id: I6c5ed526382e0971e555a70946adb63a2acdf180
CRs-Fixed: 591767