Camera: AF circle is not coming on 1'st tap while FD is ON
ISSUE:
For ZSL mode, face detection last callback is received in 40-50ms
after calling stopfacedetection on camera device. Due to this,
even though application immediately clears UI, invalid circles
remain on screen. To avoid that, earlier fix made to delay clearing
of face circles by 100ms. With that fix, for non-zsl case, AF
circle is not appearing on 1'st tap with FD.

Fix:
Change made in HAL so that this delay issue will not be there.
In application, made the code aligned to base android by removing
the 100ms delay in clearing faces.

Change-Id: I44b4d5c40c45aad553a00243f9b39f414b8e3c1e
CRs-Fixed: 638111