Reorder flash setting array entry values.
In the flashmode entry values array, torch value was at 0th position
and off at 1st postion. So when user set flash mode to off, element
at index 1 was chosen. But custom checkbox logic enables or disables
checkbox based on index of array elements. Because of this checkbox
was representing reverse values. So changing the order of array
entry values.

Change-Id: If9ae3d73b479de34fee88bb5ea9e6189b9b22ba1