SnapdragonCamera: fix don't lock screen after press power key
The window flag FLAG_SHOW_WHEN_LOCKED is workding even if press power
key, so camera activity will not be close still run above locked screen
UI.

The CameraGestureActivity is start from screen off, so it will be a
SecureCameraActivity. Finish activity as same as the secure camera
activity.

Change-Id: I76d8ff3d2f037b81140b35255b52b4c4f4de95d3

CRs-Fixed: 777770