SnapdragonCamera: Fix crash when switching between camera1, camera2
Camera1 uses CameraHolder for opening/releasing camera and
it does not release right away when closing camera.
This is done to optimize opening/closing but since Camera2
does not use CameraHolder, state is not proper when switching
between camera1 and camera2.
So ignore the timer and release right away when going from
camera1 to camera2.

Change-Id: Ibc40297817ebfca04e19097bd463f78aec4d75cd
CRs-Fixed: 1019338