SnapdragonCamera: Fix can`t swap from icon for camera
RootCause:
  Design fault, When the user set lockscreen(Pattern or Password)
, Sliding camera icon from the lock screen interface, jump to the
GrantPermissionsActivity interface which lack of access to the
flag property verification. The flag in GrantPermissionsActivity`s
interface is FLAG_KEEP_SCREEN_ON.
Solution:
   Reconstruction and redesign of the process. When the user set
lock screen, camera verificate the permission which is critical
Permissions. when cameraApp has missed of any one of the three
critical Permissions, CameraApp show dialog which tell the user
missing the critical permission and finish cameraActivity. If the
cameraApp has the three critical Permissions, the user can enter
the camera from the lockScreen.

Change-Id: Ic13c376025bf4afeceb14a24ec558e9a1179a059
CRs-Fixed: 1088278