SnapdragonCamera:Fix Camera force close when touching Restore defaults.
Before restore settings,if current is not developer mode,the developer
preferenceGroup has been removed when enter camera by default.
So duplicate remove it will cause crash.

When remove developer preferenceGroup,need to check if it's null,if it's
null,indicates that it has been removed, no need to remove it again.

Change-Id: I6185011c5183ff0821d4925d9b40c077d3e354d2
CRs-Fixed: 1115107